fix booking page
